requirements.txt
# Core dependencies
PyQt5>=5.15.0
weaviate-client>=4.4.3
openai>=1.3.0
tiktoken>=0.5.1
loguru>=0.7.2
click>=8.1.7
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Data processing
numpy>=1.24.3
beautifulsoup4>=4.12.0
httpx>=0.25.1

# Type checking and validation
pydantic>=2.5.1
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0